requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.14.3
    # via -r requirements.in
certifi==2025.11.12
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via pytest
cryptography==46.0.3
    # via oracledb
exceptiongroup==1.3.1
    # via pytest
greenlet==3.1.1
    # via sqlalchemy
idna==3.11
    # via requests
iniconfig==2.1.0
    # via pytest
numpy==1.24.4
    # via pandas
oracledb==3.0.0
    # via -r requirements.in
packaging==25.0
    # via pytest
pandas==2.0.3
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
pycparser==2.23
    # via cffi
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via -r requirements.in
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2025.2
    # via pandas
requests==2.32.4
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.45
    # via -r requirements.in
tomli==2.3.0
    # via pytest
typing-extensions==4.13.2
    # via
    #   annotated-types
    #   beautifulsoup4
    #   cryptography
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.2.3
    # via requests
